impoved server files structure; removed test data; all locales were moved to the specific folder
