--- server/config.js
@@ -1,3 +1,0 @@
-module.exports = {
-	uri: 'ekb.shri14.ru/api'
-};
