+++ server/configs/urls.js
@@ -1,0 +1,3 @@
+module.exports = {
+	api: 'ekb.shri14.ru/api'
+};
