+++ server/helpers/cities-list.js
@@ -1,5 +1,5 @@
 var request = require('request'),
-	config = require('./config'),
+    urlsConfig = require('../configs/urls'),
 	url = require('url'),
 	vow = require('vow');
 
@@ -8,7 +8,7 @@ function getCities(geoid) {
 
 	var uri = url.format({
 	    protocol: 'http',
-	    hostname: config.uri,
+	    hostname: urlsConfig.api,
 	    pathname: ['localities', geoid, 'cities'].join('/')
 	});
