--- server/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en">
-<head>
-	<meta http-equiv="content-type" content="text/html; charset=utf-8">
-	<title>Title Goes Here</title>
-</head>
-<body>
-	<p>This is Server web page</p>
-</body>
-</html>
-
-
