+++ server/locales/moment.js
@@ -1,0 +1,7 @@
+module.exports = {
+    months: [
+        "января", "февраля", "марта", "апреля", "мая", "июня", "июля",
+        "августа", "сентября", "октября", "ноября", "декабря"
+    ],
+    weekdaysMin: [ 'ПН', 'ВТ', 'СР', 'ЧТ', 'ПТ', 'СБ', 'ВС' ]
+}
