+++ server/locales/timeofday.js
@@ -1,0 +1,6 @@
+module.exports = {
+    morning: 'утром',
+    day: 'днем',
+    evening: 'вечером',
+    night: 'ночью'
+};
